+++ semgrep-core/src/parsing/Parse_rule.ml
@@ -499,6 +499,7 @@ and parse_formula_old env ((key, value) : key * G.expr) : R.formula_old =
   match s with
   | "pattern" -> R.Pat (get_pattern value)
   | "pattern-not" -> R.PatNot (t, get_pattern value)
+  | "focus-metavariable" -> R.PatFocus (t, parse_string env key value)
   | "pattern-inside" -> R.PatInside (get_pattern value)
   | "pattern-not-inside" -> R.PatNotInside (t, get_pattern value)
   | "pattern-either" ->
@@ -553,7 +554,7 @@ and parse_formula_new env (x : G.expr) : R.formula =
       | "and" ->
           let xs = parse_list env key parse_formula_and_new value in
           let fs, conds = Common.partition_either (fun x -> x) xs in
-          R.And (t, fs, conds)
+          R.And (t, fs, conds, (* TODO: 'focus-metavariable:' *) [])
       | "or" -> R.Or (t, parse_list env key parse_formula_new value)
       | "not" -> R.Not (t, parse_formula_new env value)
       | "inside" -> R.P (parse_xpattern_expr env value, Some Inside)