+++ semgrep-core/tests/OTHER/rules/focus_metavariable.py
@@ -1,0 +1,5 @@
+# ruleid: test
+foo (1 + 2,
+	1)
+foo (2 + 1,
+	1)