+++ CHANGELOG.md
@@ -5,6 +5,12 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 ## Unreleased
 
 ### Fixed
+### Added
+
+- New `focus-metavariable:` operator that allows to focus matching on a particular
+  metavariable within a `patterns:` operator (#4453)
+
+## [0.83.0](https://github.com/returntocorp/semgrep/releases/tag/v0.83.0) - 2022-02-24
 
 - Show correct findings count when using `--config auto` (#4674)